+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.0.2
+
+Minor and decorative changes to some components code.
+Removed a duplicated _send method in `component-hello-world-v2-fe/templates/src/restapi.js` generator.
+Removed deprecated `.kumori` folder in generators templates.
+
 ## v1.0.1
 
 Changed the default templates used in kumoriConfig.json. The old ones were wrong.